Rename Page to Frame, Page.Page to NavbarIndicator
This makes more sense. The whole page module is really about framing the
content between header and footer. The union type previously called Page is not
representing a page per se, but rather an indicator of which link in the nav
bar should be considered active, so NavbarIndicator is much better name.
Finally, since some pages have no corresponding link in the nav bar, we now
call it None (was Other).