WIP: Reimplement navbar using Elm Bootstrap
It's a work in progress, because:

1.  The items are not visible

    It is not clear to me why. I will investigate after the weekend.

2.  The code is somewhat spaghetti like

    With 5-argument functions and such. I hope to refactor it once I got the
    basic functionality.

What I did so far is:

Turn Main.Model into a record, with the union type previously called Model
renamed to Page and tracked in model.page field. There is also another field
called navbar with Bootstrap.Navbar.State type. It keeps the state of the
navbar. Helper functions had to be adjusted to this fundamental change and some
of them have pretty weird signatures ATM. Generally in my experience it's never
good idea to make Main.Model a union type. I tried it few times and it always
backfires as soon as the app needs to keep some additional piece of information
that is shared between different states. I'm surprised that Richard Feldman
modeled it that way.

The init returns the initial navbar command and the Navbar.subscriptions is
always there.

The biggest changes are in `Frame` module which is responsible for wrapping
content of each page with header and footer. The navbar is in the header and
needs to get it's config (including `Main.Msg` constructor) and state from
`Main`. So the changes propageted to `Main.view` which calls the aforementioned
and need to pass these additional parameters now. My feeling at this time is
that separating `Frame` from the `Main` module complicates things
unnecessarily. Perhaps on Monday I'll try to merge them together and see how it
will work. But I don't think the main issue (items not being visible) is
related to this.