Limit pagination to 5 neighboring pages
It's a quick fix to prevent pagination links from expanding the width of the
page. In the original SPA they are wrapping (display: inline), but IMO it looks
ugly. So instead of recreating the style I decided to add a little bit of logic
and get a 5-pages span around the current page. With a little more effort it
should be easy to add [ first ], [ previous ], [ next ] and [ last ] buttons,
but for now I want to cover the basics and this seems good enough. There are
bigger issues around pagination anyway:

- It is not preserved when navigating back and forth. Navigate to page 2, then
  to one of the articles and back - you will get to page 1 instead of 2.

- When navigating to a different page the pagination links are updated
  immediately, but the page content stays the same for fraction of a second
  before the HTTP response comes and gets decoded. It makes a weird UX where
  the navigation seems broken and user (me?) wants to click again. But then new
  content appears and click goes to some random element 😬.

These two problems are not spacific to the Elm implementation. The React -
Redux (most popular one) has them too.