Make banner on article page dark and responsive
The responsivness is not perfect - particularly on small screens there is no
vertical space between buttons. I hesitate to change the definition of the
buttons, as they are reused in multiple places. Also I looked at the upstream
Elm implementation and React - Redux one and they are far worse in this respect
so I will pause here and call it decent. We can address it in a separate PR.